Split adaptors/liteAdaptor.ts into separate files for each class (making them more reusable).
